Update change_wdir, delete_dir, list, and print commands.

+++ multics/filesystem/sss/commands/delete_dir.py
@@ -1,10 +1,13 @@
 
 from multics.globals import *
 
+include.query_info
+
 def delete_dir():
     declare (get_wdir_ = entry . returns (char(168)))
     arg_list    = parm()
     dir_to_kill = parm()
+    answer      = parm()
     code        = parm()
     
     current_dir = get_wdir_()
@@ -20,10 +23,15 @@ def delete_dir():
         return
     # end if
     
-    # call.ioa_("Delete directory {0}", dir_to_kill.name)
-    call.hcs_.delete_branch_(dir_to_kill.name, code)
-    if code.val != 0:
-        call.ioa_("Could not delete directory")
+    query_info.version = query_info_version_5
+    query_info.suppress_name_sw = False
+    query_info.suppress_spacing = True
+    query_info.yes_or_no_sw = True
+    call.command_query_(query_info, answer, "delete_dir", "Do you want to delete the directory\n{0:12}{1}??  ", "", dir_to_kill.name)
+    if answer.val.lower() in ["yes", "y"]:
+        call.hcs_.delete_branch_(dir_to_kill.name, code)
+        if code.val != 0:
+            call.ioa_("Could not delete directory")
         
 #-- end def delete_dir
